style: Added "Modern" Logo
Flathub insists that logos should be "in line with contemporary styles",
whatever that means. Anyway, on the Flathub listing for Bforartists, the
existing icon doesn't fit this requirement, so to make things more
"contemporary", I've took the existing icon and modified it to be a flat
color rather then the slightly detailed Bforartists logo.